+++ packages/generator/src/v1.1/generateClientSdkVersionOneDotOne.ts
@@ -19,10 +19,12 @@ import type { MinimalFs } from "../MinimalFs";
 import { sanitizeMetadata } from "../shared/sanitizeMetadata";
 import type { WireOntologyDefinition } from "../WireOntologyDefinition";
 import { generateActions } from "./generateActions";
+import { generateBackCompatDeprecatedExports } from "./generateBackCompatDeprecatedExports";
 import { generateFoundryClientFile } from "./generateFoundryClientFile";
 import { generateIndexFile } from "./generateIndexFile";
 import { generateMetadataFile } from "./generateMetadataFile";
 import { generateObjectsInterfaceFile } from "./generateObjectsInterfaceFile";
+import { generateObjectsInterfaceSupportFiles } from "./generateObjectsInterfaceSupportFiles";
 import { generateOntologyIndexFile } from "./generateOntologyIndexFile";
 import { generatePerActionDataFiles } from "./generatePerActionDataFiles";
 import { generatePerObjectInterfaceAndDataFiles } from "./generatePerObjectInterfaceAndDataFiles.1";
@@ -43,6 +45,11 @@ export async function generateClientSdkVersionOneDotOne(
   await generateMetadataFile(sanitizedOntology, fs, outDir);
   await generateOntologyIndexFile(fs, path.join(outDir, "ontology"));
   await generateObjectsInterfaceFile(sanitizedOntology, fs, objectsDir);
+  await generateObjectsInterfaceSupportFiles(
+    sanitizedOntology,
+    fs,
+    path.join(objectsDir, "objects-api"),
+  );
   await generatePerObjectInterfaceAndDataFiles(
     sanitizedOntology,
     fs,
@@ -61,4 +68,5 @@ export async function generateClientSdkVersionOneDotOne(
     queriesDir,
   );
   await generateIndexFile(fs, outDir);
+  await generateBackCompatDeprecatedExports(fs, outDir);
 }