+++ examples/todoapp/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @osdk/examples.todoapp
 
+## 0.0.6
+
+### Patch Changes
+
+- 23a3515: Add deprecation notice to objects-api types
+- Updated dependencies [23a3515]
+  - @osdk/legacy-client@0.0.15
+  - @osdk/api@0.0.16
+
+## 0.0.5
+
+### Patch Changes
+
+- 4f73749: Re-add the types to support the ObjectSet APIs
+- Updated dependencies [4f73749]
+  - @osdk/legacy-client@0.0.14
+  - @osdk/api@0.0.15
+
+## 0.0.4
+
+### Patch Changes
+
+- ad425f9: Fix query return type for objects
+- Updated dependencies [ad425f9]
+  - @osdk/legacy-client@0.0.13
+  - @osdk/api@0.0.14
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [4f9758d]
+  - @osdk/legacy-client@0.0.12
+
+## 0.0.2
+
+### Patch Changes
+
+- 25de57b: Small fixes
+- Updated dependencies [11ac931]
+- Updated dependencies [25de57b]
+  - @osdk/api@0.0.13
+  - @osdk/legacy-client@0.0.11
+
 ## 0.0.1
 
 ### Patch Changes